Too many updates to comment on
I made way too many updates and played with a bunch of functions. But some notable things are that I made a function that can accept a cell of framestacks and a cell of axes and then saves a gif. I don't really know how good these will actually end up being, but I hope it'll be good. I also remade an advanced functween that outputs framestacks